requirements.txt
streamlit
pandas
scikit-learn
plotly
python-docx
PyMuPDF
